Clarify explanation of command chain
Excellent project, but there's an ambiguous sentence or two in the README that had me very confused. I looked at this yesterday and couldn't understand the following passage:

> Builds an image called 'dockerception' from the tar stream which contains a Dockerfile and the binary

I originally parsed this as "an image called 'dockerception' … contains a Dockerfile and the binary" rather than "the tar stream … contains a Dockerfile and the binary". Only this morning after playing with the project could I figure out that the Dockerfile and binary were part of the tar stream, not the resulting image.

I think the included patch clarifies the wording.